Update sketch.js
This current iteration draws every single pixel individually and creates a pretty cool effect, which is similar to the static effect created when a television is broken. There are probably ways to achieve this effect that are much less intensive on the processor, however the effect looks cool. Also works as a proof of concept for drawing each pixel individually.